# core/vm: less allocations for various call variants 