--- a/core/vm/interface.go
+++ b/core/vm/interface.go
@@ -58,6 +58,7 @@ type StateDB interface {
 	Empty(common.Address) bool
 
 	RevertToSnapshot(int)
+	DiscardSnapshot(int)
 	Snapshot() int
 
 	AddLog(*types.Log)