… core/vm/contracts_test.go
… core/vm/instructions.go
… core/vm/interpreter.go
… core/vm/logger.go
… core/vm/logger_json.go
… core/vm/logger_test.go
… core/vm/runtime/runtime.go
… core/vm/runtime/runtime_test.go
… core/vm/stack.go
… eth/tracers/tracer.go
… eth/tracers/tracer_test.go
… trie/secure_trie.go