This PR is a bit experimental, and contains a lot of reductions in allocations. It's based off of five different benchmarks

- Staticcall to identity precompile.
- Call to identity precompile, 
- A pretty 'normal' loop which does nothing other than loop (no calls)
- A 0-value call to an EOA (no code)
- A 0-value call to a contract which immediately reverts

In most cases, allocations could be brought down to a constant, exceptions being call to nonexist and call to reverting, where allocs scale linearly with the number of calls made (however, this PR still reduces the amount by a lot). 

```
[user@work go-ethereum]$ benchstat before.txt after.txt 
name                                   old time/op    new time/op    delta
SimpleLoop/staticcall-identity-100M-6     115ms ± 1%      62ms ± 1%   -46.25%  (p=0.008 n=5+5)
SimpleLoop/call-identity-100M-6           281ms ± 2%      81ms ± 6%   -71.28%  (p=0.008 n=5+5)
SimpleLoop/loop-100M-6                    690ms ± 4%     679ms ± 2%      ~     (p=0.548 n=5+5)
SimpleLoop/call-nonexist-100M-6           459ms ±49%     162ms ± 4%   -64.60%  (p=0.008 n=5+5)
SimpleLoop/call-EOA-100M-6                570ms ± 9%      73ms ± 1%   -87.28%  (p=0.008 n=5+5)
SimpleLoop/call-reverting-100M-6          3.30s ±60%     0.17s ± 1%   -94.90%  (p=0.016 n=5+4)

name                                   old alloc/op   new alloc/op   delta
SimpleLoop/staticcall-identity-100M-6    42.9MB ± 0%     0.0MB ± 2%   -99.99%  (p=0.008 n=5+5)
SimpleLoop/call-identity-100M-6          42.7MB ± 0%     0.0MB ± 2%   -99.99%  (p=0.008 n=5+5)
SimpleLoop/loop-100M-6                   14.0kB ± 0%     4.2kB ±18%   -70.28%  (p=0.008 n=5+5)
SimpleLoop/call-nonexist-100M-6           105MB ± 0%      15MB ± 0%   -85.41%  (p=0.008 n=5+5)
SimpleLoop/call-EOA-100M-6               43.6MB ± 0%     0.0MB ± 0%   -99.99%  (p=0.016 n=5+4)
SimpleLoop/call-reverting-100M-6         1.24GB ± 0%    0.06GB ± 0%   -95.54%  (p=0.008 n=5+5)

name                                   old allocs/op  new allocs/op  delta
SimpleLoop/staticcall-identity-100M-6      670k ± 0%        0k ± 0%   -99.99%  (p=0.008 n=5+5)
SimpleLoop/call-identity-100M-6            668k ± 0%        0k ± 0%   -99.99%  (p=0.008 n=5+5)
SimpleLoop/loop-100M-6                     42.0 ± 0%      31.0 ± 6%   -26.19%  (p=0.008 n=5+5)
SimpleLoop/call-nonexist-100M-6           1.09M ± 0%     0.27M ± 0%   -75.00%  (p=0.008 n=5+5)
SimpleLoop/call-EOA-100M-6                 681k ± 0%        0k ± 0%  -100.00%  (p=0.008 n=5+5)
SimpleLoop/call-reverting-100M-6          1.62M ± 0%     1.08M ± 0%   -33.34%  (p=0.008 n=5+5)
```