@holiman 2020-06-14
Benchmark: https://geth-bench.ethdevops.io/d/Jpk-Be5Wk/dual-geth?orgId=1&var-exp=bench03&var-master=bench04&var-percentile=50&from=1592141282530&to=now