@holiman 2020-06-15
I took a memory dump from this PR as it was executing. 

![profile001](https://user-images.githubusercontent.com/142290/84655431-3a8f5c00-af11-11ea-8bcc-6fa98c0cb31d.png)