@holiman 2020-06-15
On the benchmark run, this PR is ahead by `35m` after ~24 hours (https://geth-bench.ethdevops.io/d/Jpk-Be5Wk/dual-geth?orgId=1&var-exp=bench03&var-master=bench04&var-percentile=50&from=1592227036577&to=1592229862284)