@holiman 2020-06-22
The bench run got stopped on a bad block -- I had kind of been expecting that: after the run was started, I added the commit `core/vm: tracer fix + staticcall gas fix `. Was curious to see how far it would get without that fix, stopped at block `9778717`. I updated to the latest docker image and started again, it continued without complains.