@holiman 2020-06-24
This is roughly one day after they've both been synced with chain tip.

![Screenshot_2020-06-24 Dual Geth - Grafana](https://user-images.githubusercontent.com/142290/85577171-4ed20800-b639-11ea-8642-01c16adc2a06.png)

And allocs (yellow is this PR) : 
![Screenshot_2020-06-24 Dual Geth - Grafana(1)](https://user-images.githubusercontent.com/142290/85577278-63ae9b80-b639-11ea-9c85-5247f3eab046.png)